requirements.txt
pysm3
astropy>=3.2.1
numpy
toml
h5py
so_pysm_models
so_noise_models
pixell
pyyaml
